pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "NetLogo_MCP"
version = "0.1.0"
description = "MCP server for NetLogo agent-based modeling"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.0.0",
    "pynetlogo>=0.5.2",
    "JPype1>=1.4.0",
    "pandas",
    "numpy",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
netlogo-mcp = "netlogo_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/netlogo_mcp"]
